refactor(bestax-migrate): share telemetry kernel with create-bestax
Same telemetry-core.ts as create-bestax (conformance fails if they
diverge). Payload builder stays here; markNoticed is gone. Core unit
tests live in create-bestax; this package covers payload shape and
reportMigrateRun.
